third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/sibling-rules-under-shadow-root-expected.txt

CONSOLE ERROR: line 87: Uncaught TypeError: element.createShadowRoot is not a function
crbug.com/337573: Test direct/indirect adjucent and positinal rules under shadow root.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
